+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # not yet released, version 0.2.0
 
 - Implemented support for `{lat, lng}` format, and corresponding function `toLatLng`.
+- Implemented conversion functions `degToRad`, `radToDeg`, `knotsToMeterPerSecond`, `meterPerSecondToKnots`, `knotsToKmPerHour`, and `kmPerHourToKnots`.
 
 
 # 2017-05-22, version 0.1.0